floating form layout with TB and helper method
the form's content is now centered using twitter bootstrap. Also creted
a few new helper methods in the application helper. These handle the
content creation of the form and take the partial as a parameter and a
optional hash. check the code for more documentation.
Addionally, the session controller now render the new action if the else
branch, in stead of the create action.