Users following now correct
Users can now follow another user. And this link is unidirectional.
Added a method 'followers' for User which returns all the users
following that particular user. This allows us to see who is following
someone, And still make sure we are not following someone just because
he decided to follow us.