TrainingSessions can now also be created
There was an issue where the current_user was not added to the trainings
session, which would result in an invisible error. The current user is
now added in the controller, but we are not yet checking if the user is
logged in.
We still need to check for authentication for all actions btw.