seperated all exercises from my exercises and created like
The Exercises tab is now a dropdown menu. My Exercises will be handeled
my the MyExercises controller, and will show the exercises you created
and the exercises you liked.
For liking I created a polimorphic model, enabling you to like exercises
and sessions.
The all Exercises tab will show all the published exercises.
I did the setup for the sessions but non of the implementation yet.
Also liking still has to be implemented.